improved index pages
