+++ server/pages14.go
@@ -5,7 +5,6 @@ import (
 	"github.com/NotLe0n/ModStats/server/helper"
 	"github.com/NotLe0n/ModStats/server/tmlapi13"
 	"github.com/NotLe0n/ModStats/server/tmlapi14"
-	"github.com/frustra/bbcode"
 	"github.com/gin-gonic/gin"
 	"html/template"
 	"net/http"
@@ -29,7 +28,7 @@ func indexPage14(c *gin.Context) {
 		"combined":  combinedDownloads(ModList),
 		"percent":   strconv.FormatFloat(float64(combinedDownloads(ModList[:10]))/float64(combinedDownloads(ModList))*100, 'f', 2, 64),
 		"median":    ModList[len(ModList)/2].DownloadsTotal,
-		"contribs":  80,
+		"contribs":  150,
 		"top10mods": ModList[:10],
 		"isLegacy":  false,
 	})
@@ -89,50 +88,6 @@ func modStatsPage14(c *gin.Context) {
 		modData.Tags[i].DisplayName = strings.Title(tag.DisplayName)
 	}
 
-	// set up compiler
-	bbcodeCompiler := bbcode.NewCompiler(true, true)
-	bbcodeCompiler.SetTag("size", nil)
-	bbcodeCompiler.SetTag("color", nil)
-	bbcodeCompiler.SetTag("center", nil)
-
-	for i := 1; i <= 6; i++ {
-		bbCodeToHTMLSameName(bbcodeCompiler, "h"+strconv.Itoa(i))
-	}
-
-	bbCodeToHTMLSameName(bbcodeCompiler, "u")
-	bbCodeToHTMLSameName(bbcodeCompiler, "i")
-	bbCodeToHTMLSameName(bbcodeCompiler, "hr")
-
-	bbcodeCompiler.SetTag("strike", func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = "s"
-		return out, true
-	})
-	bbcodeCompiler.SetTag("list", func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = "ul"
-		return out, true
-	})
-	bbcodeCompiler.SetTag("olist", func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = "ol"
-		return out, true
-	})
-
-	bbcodeCompiler.SetTag("spoiler", func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = "span"
-		out.Attrs = map[string]string{
-			"style": "background-color: black",
-		}
-		return out, true
-	})
-	bbcodeCompiler.SetTag("*", func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = "li"
-		return out, true
-	})
-
 	numFilledStars := int(5 * modData.VoteData.Score)
 	numEmptyStars := 5 - numFilledStars
 	stars := ""
@@ -144,6 +99,8 @@ func modStatsPage14(c *gin.Context) {
 		stars += "☆"
 	}
 
+	bbcodeCompiler := helper.NewBBCodeCompiler()
+
 	c.HTML(http.StatusOK, "base/mod.gohtml", gin.H{
 		"modlist":            tmlapi13.GetModList(),
 		"modData":            modData,
@@ -151,13 +108,5 @@ func modStatsPage14(c *gin.Context) {
 		"escapedDescription": template.HTML(helper.ParseChatTags(bbcodeCompiler.Compile(modData.Description))),
 		"stars":              stars,
 		"isLegacy":           false,
-	})
-}
-
-func bbCodeToHTMLSameName(c bbcode.Compiler, name string) {
-	c.SetTag(name, func(node *bbcode.BBCodeNode) (*bbcode.HTMLTag, bool) {
-		out := bbcode.NewHTMLTag("")
-		out.Name = name
-		return out, true
 	})
 }
